Moving the GitProvider from the core crate to the developer_platforms crate
